# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Pessoas",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/pessoas": {
            "get": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Lista de pessoas",
                "description": "Retorna uma lista de pessoas",
                "operationId": "getPessoasList",
                "responses": {
                    "200": {
                        "description": "Sucesso. Lista de pessoas retornada com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível retornar a lista de pessoas."
                    }
                }
            },
            "post": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Cadastro de pessoa",
                "description": "store",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nomePessoa",
                                    "dataNascimento",
                                    "cpf",
                                    "rg"
                                ],
                                "properties": {
                                    "nomePessoa": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Leonardo"
                                    },
                                    "dataNascimento": {
                                        "description": "Data de nascimento",
                                        "type": "string"
                                    },
                                    "cpf": {
                                        "description": "CPF",
                                        "type": "string"
                                    },
                                    "rg": {
                                        "description": "RG",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso. Cadastro realizado com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível realizar o cadastro."
                    }
                }
            }
        },
        "/api/v1/pessoas/{id}": {
            "get": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Mostra uma pessoa",
                "description": "Mostra uma pessoa",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso. Pessoa retornada com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível retornar a pessoa."
                    }
                }
            },
            "put": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Atualização de pessoa",
                "description": "update",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nomePessoa",
                                    "dataNascimento",
                                    "cpf",
                                    "rg"
                                ],
                                "properties": {
                                    "nomePessoa": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Leonardo"
                                    },
                                    "dataNascimento": {
                                        "description": "Data de nascimento",
                                        "type": "string"
                                    },
                                    "cpf": {
                                        "description": "CPF",
                                        "type": "string"
                                    },
                                    "rg": {
                                        "schema": "string",
                                        "description": "RG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso. Atualização realizada com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível realizar a atualização."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Exclusão de pessoa",
                "description": "destroy",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso. Exclusão realizada com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível realizar a exclusão."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport apiKey security.",
                "in": "header",
                "scheme": "http",
                "name": "bearer",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}